requirements.txt
Django
python-dotenv
oracledb
djangorestframework
djangorestframework-simplejwt